+++ src/commands/agent/test/run.ts
@@ -7,10 +7,16 @@
 
 import { SfCommand, Flags } from '@salesforce/sf-plugins-core';
 import { Messages, SfError } from '@salesforce/core';
-import { AgentTester } from '@salesforce/agents';
+import { AgentTester, AgentTestStartResponse } from '@salesforce/agents';
 import { colorize } from '@oclif/core/ux';
 import { CLIError } from '@oclif/core/errors';
-import { resultFormatFlag, testOutputDirFlag } from '../../../flags.js';
+import {
+  FlaggablePrompt,
+  makeFlags,
+  promptForAiEvaluationDefinitionApiName,
+  resultFormatFlag,
+  testOutputDirFlag,
+} from '../../../flags.js';
 import { AgentTestCache } from '../../../agentTestCache.js';
 import { TestStages } from '../../../testStages.js';
 import { handleTestResults } from '../../../handleTestResults.js';
@@ -24,6 +30,27 @@ export type AgentTestRunResult = {
   status: string;
 };
 
+const FLAGGABLE_PROMPTS = {
+  'api-name': {
+    char: 'n',
+    required: true,
+    message: messages.getMessage('flags.api-name.summary'),
+    validate: (d: string): boolean | string => {
+      if (d.length === 0) {
+        return true;
+      }
+      if (d.length > 80) {
+        return 'API name cannot be over 80 characters.';
+      }
+      const regex = /^[A-Za-z][A-Za-z0-9_]*[A-Za-z0-9]+$/;
+      if (!regex.test(d)) {
+        return 'Invalid API name.';
+      }
+      return true;
+    },
+  },
+} satisfies Record<string, FlaggablePrompt>;
+
 export default class AgentTestRun extends SfCommand<AgentTestRunResult> {
   public static readonly summary = messages.getMessage('summary');
   public static readonly description = messages.getMessage('description');
@@ -33,11 +60,7 @@ export default class AgentTestRun extends SfCommand<AgentTestRunResult> {
   public static readonly flags = {
     'target-org': Flags.requiredOrg(),
     'api-version': Flags.orgApiVersion(),
-    name: Flags.string({
-      char: 'n',
-      required: true,
-      summary: messages.getMessage('flags.name.summary'),
-    }),
+    ...makeFlags(FLAGGABLE_PROMPTS),
     // we want to pass `undefined` to the API
     // eslint-disable-next-line sf-plugin/flag-min-max-default
     wait: Flags.duration({
@@ -54,17 +77,31 @@ export default class AgentTestRun extends SfCommand<AgentTestRunResult> {
 
   public async run(): Promise<AgentTestRunResult> {
     const { flags } = await this.parse(AgentTestRun);
+    const connection = flags['target-org'].getConnection(flags['api-version']);
+    const apiName =
+      flags['api-name'] ?? (await promptForAiEvaluationDefinitionApiName(FLAGGABLE_PROMPTS['api-name'], connection));
 
-    this.mso = new TestStages({ title: `Agent Test Run: ${flags.name}`, jsonEnabled: this.jsonEnabled() });
+    this.mso = new TestStages({ title: `Agent Test Run: ${apiName}`, jsonEnabled: this.jsonEnabled() });
     this.mso.start();
 
-    const agentTester = new AgentTester(flags['target-org'].getConnection(flags['api-version']));
-    const response = await agentTester.start(flags.name);
+    const agentTester = new AgentTester(connection);
+    let response: AgentTestStartResponse;
+    try {
+      response = await agentTester.start(apiName);
+    } catch (e) {
+      const wrapped = SfError.wrap(e);
+      if (wrapped.message.includes('Invalid AiEvalDefinitionVersion identifier')) {
+        wrapped.actions = [
+          `Try running "sf agent test list -o ${flags['target-org'].getUsername() ?? ''}" to see available options`,
+        ];
+      }
+      throw wrapped;
+    }
 
     this.mso.update({ id: response.runId });
 
     const agentTestCache = await AgentTestCache.create();
-    await agentTestCache.createCacheEntry(response.runId, flags.name);
+    await agentTestCache.createCacheEntry(response.runId, apiName);
 
     if (flags.wait?.minutes) {
       const { completed, response: detailsResponse } = await this.mso.poll(agentTester, response.runId, flags.wait);